test: cover native and mvu_zod update paths
